yatube_api/api/serializers.py
from rest_framework import serializers
from rest_framework.relations import SlugRelatedField

from posts.models import Post, Group, Comment


class GroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = Group
        fields = ['id', 'title', 'slug', 'description']


class PostSerializer(serializers.ModelSerializer):
    author = SlugRelatedField(slug_field='username', read_only=True)
    comments = SlugRelatedField(
        many=True, required=False, slug_field='text', read_only=True
    )

    class Meta:
        model = Post
        fields = ['id', 'text', 'pub_date', 'author',
                  'image', 'group', 'comments']


class CommentSerializer(serializers.ModelSerializer):
    author = SlugRelatedField(slug_field='username', read_only=True)

    class Meta:
        model = Comment
        fields = ['id', 'author', 'post', 'text', 'created']
        read_only_fields = ('author', 'post')
